% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_seasons.R
\name{enumerate_seasons}
\alias{enumerate_seasons}
\title{Enumerates seasons in data frame}
\usage{
enumerate_seasons(data, start_month, end_month, start_day = 15,
  end_day = start_day)
}
\arguments{
\item{data}{long-format data frame with weather data; ideally output
of \code{toLong}}

\item{start_month}{numeric starting month defining season (inclusive)}

\item{end_month}{numeric ending month defining season (inclusive)}

\item{start_day}{numeric day of starting month defining season
(inclusive); defaults to 15}

\item{end_day}{numeric day of ending month defining season (inclusive);
defaults to \code{start_day}}
}
\value{
data frame with new columns, "season" and "season_year"
}
\description{
Enumerates seasons in data frame
}
