% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_summarize_rainfall.R
\name{par_summarize_rainfall}
\alias{par_summarize_rainfall}
\title{A wrapper for \code{summarize_rainfall} to perform summary statistics
calculations in parallel}
\usage{
par_summarize_rainfall(num_cores = NULL, ...)
}
\arguments{
\item{num_cores}{integer indicating number of processors to use; if
\code{NULL} (default), uses one fewer than the number of processors available}

\item{...}{additional values passed to \code{\link{summarize_rainfall}}}
}
\value{
tibble with rainfall summary statistics
}
\description{
A wrapper for \code{summarize_rainfall} to perform summary statistics
calculations in parallel
}
\seealso{
\code{\link{summarize_rainfall}}
}
