% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain_2yr}
\alias{rain_2yr}
\title{A tiny (two-year) example data set of daily rainfall measurements}
\format{
An object of class \code{data.frame} with 6 rows and 732 columns.
}
\usage{
rain_2yr
}
\description{
The data set includes one column with the unique identifier for the site
(\code{y4_hhid}) and remaining 731 columns with daily rainfall measurements.
Columns are named with the "rf_YYYYMMDD" format. For example, the column
\code{rf_19830101} has rainfall values for 1983-01-01.
}
\keyword{datasets}
