% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rain_4yr}
\alias{rain_4yr}
\title{A small (four-year) example data set of daily rainfall measurements}
\format{
An object of class \code{data.frame} with 100 rows and 1462 columns.
}
\usage{
rain_4yr
}
\description{
The data set includes one column with the unique identifier for the site
(\code{y4_hhid}) and remaining 1,461 columns with daily rainfall
measurements. Columns are named with the "rf_YYYYMMDD" format. For example,
the column \code{rf_19830101} has rainfall values for 1983-01-01.
}
\keyword{datasets}
