% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_seasons.R
\name{season_year}
\alias{season_year}
\title{Value for season year for observation}
\usage{
season_year(x, start_month, end_month, start_day = 15,
  end_day = start_day)
}
\arguments{
\item{x}{vector of Date data}

\item{start_month}{numeric starting month defining season (inclusive)}

\item{end_month}{numeric ending month defining season (inclusive)}

\item{start_day}{numeric day of starting month defining season
(inclusive); defaults to 15}

\item{end_day}{numeric day of ending month defining season (inclusive);
defaults to \code{start_day}}
}
\value{
integer vector of season year to which observation corresponds to
}
\description{
Value for season year for observation
}
