% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_2yr}
\alias{temperature_2yr}
\title{A tiny (two-year) example data set of daily temperature measurements}
\format{
An object of class \code{data.frame} with 6 rows and 732 columns.
}
\usage{
temperature_2yr
}
\description{
The data set includes one column with the unique identifier for the site
(\code{hhid}) and remaining 731 columns with daily temperature measurements
in degrees Celsius. Columns are named with the "tmp_YYYYMMDD" format. For
example, the column \code{tmp_19830101} has temperature values for
1983-01-01.
}
\keyword{datasets}
