% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temperature_4yr}
\alias{temperature_4yr}
\title{A small (four-year) example data set of daily temperature measurements}
\format{
An object of class \code{data.frame} with 100 rows and 1462 columns.
}
\usage{
temperature_4yr
}
\description{
The data set includes one column with the unique identifier for the site
(\code{hhid}) and remaining 1,461 columns with daily temperature measurements
in degrees Celsius. Columns are named with the "tmp_YYYYMMDD" format. For
example, the column \code{tmp_19830101} has temperature values for
1983-01-01.
}
\keyword{datasets}
