% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_long.R
\name{to_long}
\alias{to_long}
\title{Convert wide format weather data to long format}
\usage{
to_long(data, keep_cols = 1, date_sep = "_")
}
\arguments{
\item{data}{wide format data}

\item{keep_cols}{columns in \code{data} to retain in long format; all other 
columns will be converted to key/value pairs in long format; defaults to 1}

\item{date_sep}{character used to delimit variable prefix from date in 
column names; defaults to underscore ("_")}
}
\value{
data frame where each row corresponds to an observation for a single 
day
}
\description{
Convert wide format weather data to long format
}
