% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_summary.R
\name{wide_summary}
\alias{wide_summary}
\title{Wide-formatted summary data}
\usage{
wide_summary(x, id_col, year_col = "year", long_term_cols = "")
}
\arguments{
\item{x}{summary data in pseudo-wide format}

\item{id_col}{name of column with id value}

\item{year_col}{name of column with year value}

\item{long_term_cols}{vector of names of columns not to be transformed 
to separate values for each year}
}
\value{
wide-formatted data, where summary statistics for individual years 
are in separate columns
}
\description{
Wide-formatted summary data
}
