% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_interval.R
\name{dry_interval}
\alias{dry_interval}
\title{Longest stretch of consecutive dry days}
\usage{
dry_interval(x, rain_cutoff = 1, period = c("start", "mid", "end"),
  na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector of rainfall measurements}

\item{rain_cutoff}{minimum amount of rainfall to count as non-dry day}

\item{period}{period to measure longest dry spell; see \code{return}}

\item{na.rm}{logical indicating treatment of \code{NA} values; by 
default, \code{na.rm = TRUE}, \code{NA} values are ignored. If set to 
\code{FALSE}, any \code{NA} values in \code{x} will result in a return value 
of \code{NA}}
}
\value{
numeric vector of length 1 with the number of days that have rainfall
less than \code{rain_cutoff}; value returned is determined by value of
\code{period}:
\describe{
  \item{start}{number of consecutive days at beginning of season with less
  than \code{rain_cutoff} of measured rain; if first day of season had
  rainfall greater than or equal to \code{rain_cutoff}, the returned value
  will be zero}
  \item{mid}{longest stretch of days with less than \code{rain_cutoff}
  contained within the period; if rainfall was less than \code{rain_cutoff}
  for every day in defined season, the returned value will be zero}
  \item{end}{number of consecutive days at end of season with less
  than \code{rain_cutoff} of measured rain; if last day of season had
  rainfall greater than or equal to \code{rain_cutoff}, the returned value
  will be zero}
}
}
\description{
Longest stretch of consecutive dry days
}
