% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dry_interval.R
\name{dry_interval}
\alias{dry_interval}
\title{Longest interval of consecutive dry days}
\usage{
dry_interval(
  x,
  rain_cutoff = 1,
  period = c("start", "mid", "end"),
  na.rm = TRUE
)
}
\arguments{
\item{x}{numeric vector of rainfall measurements}

\item{rain_cutoff}{minimum amount of rainfall to count as non-dry day}

\item{period}{period to measure longest dry spell; see \code{value}}

\item{na.rm}{logical indicating treatment of \code{NA} values; by
default, \code{na.rm = TRUE}, \code{NA} values are ignored. If set to
\code{FALSE}, any \code{NA} values in \code{x} will result in a return value
of \code{NA}}
}
\value{
numeric vector of length 1 with the number of days that have rainfall
less than \code{rain_cutoff}; value returned is determined by value of
\code{period}:
\describe{
  \item{start}{number of consecutive days at beginning of season with less
  than \code{rain_cutoff} of measured rain; if first day of season had
  rainfall greater than or equal to \code{rain_cutoff}, the returned value
  will be zero}
  \item{mid}{longest stretch of days with less than \code{rain_cutoff}
  contained within the period; if rainfall was less than \code{rain_cutoff}
  for every day in defined season, the returned value will be zero}
  \item{end}{number of consecutive days at end of season with less
  than \code{rain_cutoff} of measured rain; if last day of season had
  rainfall greater than or equal to \code{rain_cutoff}, the returned value
  will be zero}
}
}
\description{
Determines the longest interval of consecutive days that are
considered "dry". Values of \code{x} that are less than \code{rain_cutoff}
are categorized as dry.
}
\details{
Missing values are not simply removed for purpose of calculation.
If \code{x} includes \code{NA} values and \code{na.rm = FALSE}, \code{NA} is
returned. However, when \code{x} includes \code{NA} values and
\code{na.rm = TRUE}, special consideration is required. In contrast with
other methods, \code{NA} cannot be dropped without significant implications
on calculations. Consider a vector of \code{c(0, 0, NA, 0)}. If \code{NA}
values are removed, the vector reduces to \code{c(0, 0, 0)} and the longest
consectutive stretch of dry days becomes 3. This is not an accurate
representation of the data. Instead, when \code{na.rm = TRUE}, vectors with
\code{NA} values will evaluate missing values as non-dry days. In the
example above, this results in the longest dry interval being 2. This
provides a more accurate representation of dry day intervals based on
non-missing data.
}
\examples{
\dontrun{
  rain <- c(0, 2, 3, 0, 0, 2, 0, 3, 0, 0, 0)
  # Longest interval at the end of season
  dry_test <- wxsumR:::dry_interval(x = rain, period = "end")
  dry_test
  # 3

  # Longest interval in middle of season
  dry_test <- wxsumR:::dry_interval(x = rain, period = "mid")
  dry_test
  # 2

  # Longest interval at start of season, considering days with rain of 2 or
  # lower (unitless) as dry days
  dry_test <- wxsumR:::dry_interval(x = rain, rain_cutoff = 2, period = "start")
  dry_test
  # 2
}
}
