% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate_seasons.R
\name{enumerate_seasons}
\alias{enumerate_seasons}
\title{Enumerate seasons in data frame}
\usage{
enumerate_seasons(
  data,
  start_month,
  end_month,
  start_day = 15,
  end_day = start_day
)
}
\arguments{
\item{data}{long-format data frame with weather data including a
column \code{date}; ideally the data object is the result of a call to
\code{\link{to_long}}}

\item{start_month}{numeric starting month defining season (inclusive)}

\item{end_month}{numeric ending month defining season (inclusive)}

\item{start_day}{numeric day of starting month defining season
(inclusive); defaults to 15}

\item{end_day}{numeric day of ending month defining season (inclusive);
defaults to \code{start_day}}
}
\value{
A copy of the passed \code{data} data frame with an additional
column, \code{season_year} indicating the year of the season that the
observation "belongs" to. For seasons that \emph{do not} cross the new year,
this is simply the year the observation was made. For seasons that \emph{do}
include the new year (e.g. a season starts in November and ends in the
subsequent March), \code{season_year} is defined as the year in which the
season starts. That is, for a season spanning November through March, the
value of \code{season_year} for observations in January, February, and March
is the prior year (e.g. \code{season_year} for an observation on 1983-02-01
for a season spanning November through March would be 1984).
}
\description{
\code{enumerate_seasons} categorizes rows by an annual season
defined by \code{start_*} and \code{end_*} parameters. The resulting object
includes a \code{season_year} column indicating which season the date is
considered part of. This is generally only used for internal data
processing by \code{summarize_*} functions.
}
\examples{
\dontrun{
# to_long and enumerate_seasons are internal functions, hence the triple-
colon (:::)
# enumerate_seasons expects long-formatted data
rain_long <- weathercommand:::to_long(data = rain_2yr)

# enumerate seasons, defined by 30 November through 15 March
rain_seasons <- weathercommand:::enumerate_seasons(data = rain_long,
                                                   start_month = 11,
                                                   end_month = 3,
                                                   start_day = 30,
                                                   end_day = 15)
}

}
