% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_summarize_rainfall.R
\name{par_summarize_rainfall}
\alias{par_summarize_rainfall}
\title{Rainfall summary statistics calculated in parallel}
\usage{
par_summarize_rainfall(rain, num_cores = NULL, id_index = 1, ...)
}
\arguments{
\item{rain}{data frame with daily rainfall data for each site}

\item{num_cores}{integer indicating number of processors to use; if
\code{NULL} (default), uses one fewer than the number of processors available}

\item{id_index}{integer column index of unique site id}

\item{...}{additional values passed to
\code{\link{summarize_rainfall}}}
}
\value{
tibble with rainfall summary statistics
}
\description{
A wrapper for \link{summarize_rainfall} to perform rainfall
summary statistics calculations in parallel.
}
\examples{
\dontrun{
# Season defined by 15 March through 15 November
rain_summary <- par_summarize_rainfall(rain = rain_2yr,
                                       start_month = 3,
                                       end_month = 11)

# Same as example above, but restrict use to 2 processors
rain_summary <- par_summarize_rainfall(rain = rain_2yr,
                                       start_month = 3,
                                       end_month = 11,
                                       num_cores = 2)

# Season defined by 15 March through 15 November, output separate rows for
# each year
rain_summary <- par_summarize_rainfall(rain = rain_2yr,
                                       start_month = 3,
                                       end_month = 11,
                                       wide = FALSE)
}

}
\seealso{
\code{\link{summarize_rainfall}}
}
