% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_summarize_temperature.R
\name{par_summarize_temperature}
\alias{par_summarize_temperature}
\title{A wrapper for \code{summarize_temperature} to perform summary statistics
calculations in parallel}
\usage{
par_summarize_temperature(temperature, num_cores = NULL, id_index = 1,
  ...)
}
\arguments{
\item{temperature}{data frame with daily temperature data for each site}

\item{num_cores}{integer indicating number of processors to use; if
\code{NULL} (default), uses one fewer than the number of processors available}

\item{id_index}{integer column index of unique site id}

\item{...}{additional values passed to \code{\link{summarize_temperature}}}
}
\value{
tibble with temperature summary statistics
}
\description{
A wrapper for \code{summarize_temperature} to perform summary statistics
calculations in parallel
}
\seealso{
\code{\link{summarize_temperature}}
}
