% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_rainfall.R
\name{summarize_rainfall}
\alias{summarize_rainfall}
\title{Provides rainfall summary statistics}
\usage{
summarize_rainfall(rain, start_month, end_month, start_day = 15,
  end_day = start_day, rain_cutoff = 1, na.rm = TRUE, wide = TRUE)
}
\arguments{
\item{rain}{data frame with daily rainfall data for each site}

\item{start_month}{numeric starting month defining season (inclusive)}

\item{end_month}{numeric ending month defining season (inclusive)}

\item{start_day}{numeric day of starting month defining season
(inclusive); defaults to 15}

\item{end_day}{numeric day of ending month defining season (inclusive);
defaults to \code{start_day}}

\item{rain_cutoff}{numeric minimum value for daily rainfall to be counted as
a rain day}

\item{na.rm}{logical passed to summary statistic functions indicating 
treatment of \code{NA} values}

\item{wide}{logical indicating whether or not to output as wide-
formatted data}
}
\value{
data frame with rainfall summary statistics
}
\description{
Provides rainfall summary statistics
}
\seealso{
\code{\link{summarize_temperature}}
}
