% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_temperature.R
\name{summarize_temperature}
\alias{summarize_temperature}
\title{Provides temperature summary statistics}
\usage{
summarize_temperature(
  temperature,
  start_month,
  end_month,
  start_day = 15,
  end_day = start_day,
  growbase_low = 10,
  growbase_high = 20,
  na.rm = TRUE,
  wide = TRUE
)
}
\arguments{
\item{temperature}{data frame with daily temperature data for each site}

\item{start_month}{numeric starting month defining season (inclusive)}

\item{end_month}{numeric ending month defining season (inclusive)}

\item{start_day}{numeric day of starting month defining season
(inclusive)}

\item{end_day}{numeric day of ending month defining season (inclusive);
defaults to \code{start_day}}

\item{growbase_low}{numeric lower bound for calculating growing degree days
(inclusive)}

\item{growbase_high}{numeric upper bound for calculating growing degree days
(inclusive)}

\item{na.rm}{logical passed to summary statistic functions indicating 
treatment of \code{NA} values}

\item{wide}{logical indicating whether or not to output as wide-
formatted data}
}
\value{
data frame with temperature summary statistics
}
\description{
Provides temperature summary statistics
}
\seealso{
\code{\link{summarize_rainfall}}
}
